Bump version of seed
and remove code duplicated code
